jkjkj

+++ GreenSell/WebContent/WEB-INF/view/sell/sell_detail.jsp
@@ -20,7 +20,12 @@
 		});
 		
 		$("#deleteitem").click(function(){
+			if(confirm('정말로 삭제하시겠습니까? \r\n 두번 안 묻습니다.'))
+			{
 			location.href="deleteitem?no=${itemone.no}";
+			}else{
+				return;
+				}
 		});
 	});
